# docs/api/openapi_plp_v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SHIOL+ Powerball Prediction API",
    "description": "Provides ML-based Powerball number predictions.",
    "version": "6.0.0"
  },
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List all users",
        "description": "Returns a list of all users with basic info. Admin only.\n- id: User ID\n- username: Username\n- email: Email\n- is_admin: Admin status\n- premium_until: Premium expiration\n- created_at: Account creation date",
        "operationId": "list_users_api_v1_admin_users_get",
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "items": {},
                  "type": "array",
                  "title": "Response List Users Api V1 Admin Users Get"
                }
              }
            }
          },
          "403": {
            "description": "Admin required"
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/reset-password": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Reset user password",
        "description": "Resets the password for a user and returns a temporary password. Admin only.\nLogs the action for audit.",
        "operationId": "reset_user_password_api_v1_admin_users__user_id__reset_password_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset successful",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Password reset failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete user account",
        "description": "Deletes a user account. Admin only. Prevents self-deletion.\nLogs the action for audit.",
        "operationId": "delete_user_api_v1_admin_users__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "Admin cannot delete own account"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "User deletion failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/premium": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Toggle user premium status",
        "description": "Toggles premium status for a user. If not premium, assigns 30 days; if premium, removes. Admin only.\nLogs the action for audit.",
        "operationId": "toggle_premium_api_v1_admin_users__user_id__premium_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Premium status toggled",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/pipeline/execution-logs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get pipeline execution logs",
        "description": "Returns pipeline execution logs with optional filters. Admin only.\n\nQuery parameters:\n- limit: Maximum number of logs to return (default: 20, max: 100)\n- status: Filter by status ('running', 'completed', 'failed', 'timeout')\n- start_date: Filter logs after this date (YYYY-MM-DD)\n- end_date: Filter logs before this date (YYYY-MM-DD)\n\nReturns:\n- logs: Array of execution records sorted by start_time DESC\n- statistics: Summary statistics (total runs, success rate, avg duration, etc.)",
        "operationId": "get_pipeline_logs_api_v1_admin_pipeline_execution_logs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pipeline execution logs with statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Pipeline Logs Api V1 Admin Pipeline Execution Logs Get"
                }
              }
            }
          },
          "403": {
            "description": "Admin required"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/pipeline/trigger": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Trigger full pipeline run",
        "description": "Triggers the full pipeline execution. Admin only.\n\nParams:\n- async_run: If True (default), returns immediately after scheduling the run.\n             If False, waits for the pipeline to finish and returns the result.",
        "operationId": "trigger_pipeline_api_v1_admin_pipeline_trigger_post",
        "parameters": [
          {
            "name": "async_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Async Run"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pipeline started",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "Admin required"
          },
          "500": {
            "description": "Failed to start pipeline"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}